gralloc: add framebuffer blank / unblank calls
Many devices require a FBIOBLANK:FB_BLANK_UNBLANK ioctl be sent to the
kernel framebuffer device, before enabling the display driver, unless
it was already initialized by the bootloader.

However, the default gralloc/framebuffer or hwcomposer modules don't
do this.

- AOSP recovery fixed this problem here:
https://android.googlesource.com/platform/bootable/recovery/+/f6abd409bb2b80dfa0c1c03c26bc83141f52c673
- Jetson's 3.10 kernel requires this (normally sent via Nvidia's hwcomposer
in case of HDMI)
- Any 3.4+ kernel from omapzoom also gets this in their hwcomposer:
http://omapzoom.org/?p=platform/hardware/ti/omap4xxx.git;a=blob;f=hwc/display.c;h=0f2efc560eface4abfa45d528d1209ab5fff4c88;hb=refs/heads/d-kitkat-mr2.1-release#l691

Let's add this functionality to the default framebuffer module as well.

Change-Id: I6cf491bf91f88c79c7b6f668ecabf3b8ae73cd3e
Signed-off-by: Michael Scott <michael.scott@linaro.org>